requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
llvmlite==0.43.0
lxml==5.3.0
numba==0.60.0
numpy==1.26.4
pandas==2.2.3
pandas-datareader==0.10.0
pandas_ta @ git+https://github.com/twopirllc/pandas-ta.git@438f3a97b71b26bc17efb1a204fb0e0fdc48e0aa
pybit==5.8.0
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
